fix(summarizer): resolve pr-verify findings incl. critical claim bug
Critical:
- LawChangeRow.RetryCount int -> long. SQLite INTEGER materializes as
  Int64; Dapper's positional-record match is exact, so ClaimNextAsync threw
  on every pending row and the worker summarized nothing. Reproduced live,
  now claims + processes correctly.

Medium:
- Migrations fails fast with a clear message when law_changes is absent,
  instead of a cryptic "no such table" crash from ALTER (Program wraps it).
- Worker guards the whole item so a repo failure during failure-handling
  can't fault the BackgroundService.

Low:
- MarkSuccess UPDATE + cost INSERT now run in one transaction (atomic).
- MarkFallback clears summary so no stale summary persists on a fallback row.
- Review gate counts only successful summaries (summary IS NOT NULL), so
  fallback rows don't consume "first-N" slots.
- Fail fast when no Claude API key is configured.
- Drop obsolete `version` key from docker-compose.

+++ src/summarizer/NomadRules.Summarizer/Summarization/Models.cs
@@ -17,8 +17,8 @@ public class ClaudeOptions
     public int TimeoutSeconds { get; set; } = 30;
 }
 
-// A row claimed for summarization.
-public record LawChangeRow(string Id, string RawContent, int RetryCount);
+// A row claimed for summarization. RetryCount is long to match SQLite INTEGER (Dapper requires an exact type match).
+public record LawChangeRow(string Id, string RawContent, long RetryCount);
 
 // Parsed Claude output plus usage/cost for one summary.
 public record SummaryResult(
